Make 'cp' more POSIXly correct with regards to symlink.
The POSIX norm:
If neither the -R nor -r options were specified, cp shall take actions
based on the type and contents of the file referenced by the symbolic
link, and not by the symbolic link itself.

It means that when "recurse:false", we should actually copy the content
of the file rather than the link.

Thanks to Ivan Gotovchits for his contribution.

(Closes: #1649)